// datanode-ui/src/libs/hooks/useList/useList.enums.ts
export enum UseListActionType {
  GET_ENTITY_LIST_PENDING = 'GET_ENTITY_LIST_PENDING',
  GET_ENTITY_LIST_SUCCESS = 'GET_ENTITY_LIST_SUCCESS',
  GET_ENTITY_LIST_FAILURE = 'GET_ENTITY_LIST_FAILURE',
  GET_ENTITY_PENDING = 'GET_ENTITY_PENDING',
  GET_ENTITY_SUCCESS = 'GET_ENTITY_SUCCESS',
  GET_ENTITY_FAILURE = 'GET_ENTITY_FAILURE',
  CREATE_ENTITY_PENDING = 'CREATE_ENTITY_PENDING',
  CREATE_ENTITY_SUCCESS = 'CREATE_ENTITY_SUCCESS',
  CREATE_ENTITY_FAILURE = 'CREATE_ENTITY_FAILURE',
  UPDATE_ENTITY_PENDING = 'UPDATE_ENTITY_PENDING',
  UPDATE_ENTITY_SUCCESS = 'UPDATE_ENTITY_SUCCESS',
  UPDATE_ENTITY_FAILURE = 'UPDATE_ENTITY_FAILURE',
  DELETE_ENTITY_PENDING = 'DELETE_ENTITY_PENDING',
  DELETE_ENTITY_SUCCESS = 'DELETE_ENTITY_SUCCESS',
  DELETE_ENTITY_FAILURE = 'DELETE_ENTITY_FAILURE',
  GET_ALLOWED_SORTING_SUCCESS = 'GET_ALLOWED_SORTING_SUCCESS',
}